Sorting test, Front-end fix for checking 'GO BACK'

+++ src/main/java/com/memento/web/domain/UserRepository.java
@@ -11,7 +11,7 @@ public interface UserRepository extends MongoRepository<User, String> {
     Optional<User> findByName(String name);
 
     @Query(value = "{'name': ?0}", fields = "{'historyList': {$slice: [?1, ?2]}}")
-    Optional<User> findAllByNameWithPagination(String name, int skip, int limit);
+    Optional<User> findAllByNameWithPagination(String name, long skip, int limit);
 
 //    @Query(value = "{'name': ?0, 'historyList': { $elemMatch: {'keyword': {$regex: ?1, $options:'i'}}}", fields = "{'historyList': {$slice: [?2, ?3]}}")
 }
